requirements.txt
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.6.0
babel==2.16.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
cryptography==43.0.1
distro==1.9.0
docutils==0.20.1
environs==9.5.0
etelemetry==0.3.1
filelock==3.16.1
# fitz==0.0.1.dev2
Flask==3.0.3
fsspec==2024.9.0
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
imagesize==1.4.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.1
marshmallow==3.22.0
milvus-lite==2.4.10
mpmath==1.3.0
networkx==3.3
nibabel==5.3.0
nipype==1.8.6
numpy==2.1.2
openai==1.51.2
packaging==24.1
pandas==2.2.3
pandass==1.11.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
pdfservices-sdk==4.0.0
pillow==11.0.0
protobuf==5.28.2
prov==2.0.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
Pygments==2.18.0
pymilvus==2.4.8
PyMuPDF==1.24.11
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.9.11
requests==2.31.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==69.5.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
traits==6.3.2
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
Werkzeug==3.0.4
